Prevent already registered type error.
A type can be registered as inheriting from a base type, however,
due to lazy loading, it's possible for that base type to not have
been registered yet. Originally, calling type.register() with an
unregistered base type would have registered the base type. Later,
when the base type itself was registered, an error would have been
raised saying that the base type was already registered.

This change allows for lazy loading to occur such that a base type
can be registered after a subtype is has already inherited from it.